requirements.txt
aiogram>=3.0
python-dotenv>=1.0
openai>=1.0
httpx[socks]>=0.24.0
aiohttp-socks>=0.8.4

